# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SEEKH",
    "description": "This is a sample Course Managment Server based on the OpenAPI 3.0 specification.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/user/{id}": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updates user info",
        "description": "Name, email and password update",
        "operationId": "userUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "description": "send name email password to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user info",
        "description": "Get name and email of user",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/{request}/{courseId}/{userId}": {
      "post": {
        "tags": [
          "enrollment"
        ],
        "summary": "Accept or Reject an enrollment request",
        "description": "Accept or Reject an enrollment request",
        "operationId": "request",
        "parameters": [
          {
            "name": "request",
            "in": "path",
            "required": true,
            "type": "string",
            "enum": [
              "accept",
              "reject"
            ]
          },
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Enrollment request Approved OR Enrollment request Rejected."
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "jwt_token",
        "in": "header"
      }
    }
  }
}